+++ test/unittests/skills/test_skill_updater.py
@@ -13,29 +13,116 @@
 # limitations under the License.
 #
 """Unit tests for the SkillUpdater class."""
+import tempfile
 from os import path
-from time import sleep
+from shutil import rmtree
+from time import sleep, time
+from unittest import TestCase
 from unittest.mock import Mock, patch, PropertyMock
 
+from msm import MycroftSkillsManager
+from msm.skill_repo import SkillRepo
+from msm.util import MsmProcessLock
+from pathlib import Path
+
 from mycroft.skills.skill_updater import SkillUpdater
-from test.unittests.base import MycroftUnitTestBase
+
+MOCK_PACKAGE = 'mycroft.skills.skill_update.'
 
 
-class TestSkillUpdater(MycroftUnitTestBase):
-    mock_package = 'mycroft.skills.skill_updater.'
-    use_msm_mock = True
+class MockMessageBus:
+    def __init__(self):
+        self.message_types = []
+        self.message_data = []
+        self.event_handlers = []
 
+    def emit(self, message):
+        self.message_types.append(message.type)
+        self.message_data.append(message.data)
+
+    def on(self, event, _):
+        self.event_handlers.append(event)
+
+
+class TestSkillUpdater(TestCase):
     def setUp(self):
-        super().setUp()
+        self.message_bus = MockMessageBus()
+        self.temp_dir = tempfile.mkdtemp()
+        self._mock_msm()
+        self._mock_config()
         self._mock_time()
         self._mock_connected()
+        self._mock_dialog()
+
+    def _mock_msm(self):
+        """Define a mock object representing the MycroftSkillsManager."""
+        msm_patch = patch(MOCK_PACKAGE + 'create_msm')
+        self.addCleanup(msm_patch.stop)
+        self.create_msm_mock = msm_patch.start()
+        self.msm_mock = Mock(spec=MycroftSkillsManager)
+        self.msm_mock.skills_dir = self.temp_dir
+        self.msm_mock.platform = 'test_platform'
+        self.msm_mock.lock = MsmProcessLock()
+        self.msm_mock.repo = Mock(spec=SkillRepo)
+        self.msm_mock.repo.get_default_skill_names = Mock(
+            return_value=[
+                ('default', ['time', 'weather']),
+                ('test_platform', ['test_skill'])
+            ]
+        )
+        self.msm_mock.skills_data = dict(
+            skills=[
+                dict(name='test_skill', beta=False)
+            ]
+        )
+        skill = Mock()
+        skill.is_local = True
+        self.msm_mock.list_defaults.return_value = [skill]
+        self.create_msm_mock.return_value = self.msm_mock
+
+    def _mock_config(self):
+        """Define a mock object representing the device configuration."""
+        config_mgr_patch = patch(MOCK_PACKAGE + 'Configuration')
+        self.addCleanup(config_mgr_patch.stop)
+        self.config_mgr_mock = config_mgr_patch.start()
+        get_config_mock = Mock()
+        get_config_mock.return_value = self._build_config()
+        self.config_mgr_mock.get = get_config_mock
+
+    def _build_config(self):
+        """Build a dictionary representing device configs needed to test."""
+        return dict(
+            skills=dict(
+                msm=dict(
+                    directory='skills',
+                    versioned=True,
+                    repo=dict(
+                        cache='.skills-repo',
+                        url='https://github.com/MycroftAI/mycroft-skills',
+                        branch='19.02'
+                    )
+                ),
+                update_interval=1.0,
+                auto_update=False,
+                blacklisted_skills=[],
+                priority_skills=['foobar'],
+                upload_skill_manifest=True
+            ),
+            data_dir=self.temp_dir
+        )
 
     def _mock_connected(self):
         """Define a mock object representing the connected() function."""
-        connected_patch = patch(self.mock_package + 'connected')
+        connected_patch = patch(MOCK_PACKAGE + 'connected')
         self.addCleanup(connected_patch.stop)
         self.connected_mock = connected_patch.start()
         self.connected_mock.return_value = True
+
+    def _mock_dialog(self):
+        """Define a mock object representing the dialog module."""
+        dialog_patch = patch(MOCK_PACKAGE + 'dialog.get')
+        self.addCleanup(dialog_patch.stop)
+        self.dialog_mock = dialog_patch.start()
 
     def _mock_time(self):
         """Define a mock object representing the built-in time function.
@@ -44,25 +131,37 @@ class TestSkillUpdater(MycroftUnitTestBase):
         time.  To have the tests produce predictable results, we just need
         the time() function to return a value we can depend on.
         """
-        time_patch = patch(self.mock_package + 'time')
+        time_patch = patch(MOCK_PACKAGE + 'time')
         self.addCleanup(time_patch.stop)
         self.time_mock = time_patch.start()
         self.time_mock.return_value = 100
 
+    def tearDown(self):
+        rmtree(self.temp_dir)
+
+    def test_instantiate(self):
+        """Test the results of instantiating the class."""
+        updater = SkillUpdater(self.message_bus)
+        self.assertEqual(updater.config['data_dir'], self.temp_dir)
+        self.assertEqual(updater.update_interval, 3600)
+        self.assertEqual(
+            updater.dot_msm_path,
+            path.join(self.temp_dir, '.msm')
+        )
+        self.assertFalse(path.exists(updater.dot_msm_path))
+        self.assertLess(updater.next_download, time())
+
     def test_load_installed_skills(self):
         """Test loading a set of installed skills into an instance attribute"""
-        skill_file_path = str(self.temp_dir.joinpath('.mycroft_skills'))
+        skill_file_path = path.join(self.temp_dir, '.mycroft_skills')
         with open(skill_file_path, 'w') as skill_file:
             skill_file.write('FooSkill\n')
             skill_file.write('BarSkill\n')
 
-        patch_path = (
-            self.mock_package +
-            'SkillUpdater.installed_skills_file_path'
-        )
+        patch_path = MOCK_PACKAGE + 'SkillUpdater.installed_skills_file_path'
         with patch(patch_path, new_callable=PropertyMock) as mock_file_path:
             mock_file_path.return_value = skill_file_path
-            updater = SkillUpdater(self.message_bus_mock)
+            updater = SkillUpdater(self.message_bus)
             updater._load_installed_skills()
 
         self.assertEqual({'FooSkill', 'BarSkill'}, updater.installed_skills)
@@ -71,7 +170,7 @@ class TestSkillUpdater(MycroftUnitTestBase):
         """Test invoking MSM to install or update skills"""
         skill = self._build_mock_msm_skill_list()
         self.msm_mock.list_defaults.return_value = [skill]
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater._apply_install_or_update(quick=False)
 
         self.msm_mock.apply.assert_called_once_with(
@@ -84,7 +183,7 @@ class TestSkillUpdater(MycroftUnitTestBase):
         """Test invoking MSM to install or update skills quickly"""
         skill = self._build_mock_msm_skill_list()
         self.msm_mock.list_defaults.return_value = [skill]
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater._apply_install_or_update(quick=True)
 
         self.msm_mock.apply.assert_called_once_with(
@@ -98,7 +197,7 @@ class TestSkillUpdater(MycroftUnitTestBase):
         skill = self._build_mock_msm_skill_list()
         skill.is_local = False
         self.msm_mock.list_defaults.return_value = [skill]
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater._apply_install_or_update(quick=True)
 
         self.msm_mock.apply.assert_called_once_with(
@@ -109,14 +208,11 @@ class TestSkillUpdater(MycroftUnitTestBase):
 
     def test_save_installed_skills(self):
         """Test saving list of installed skills to a file."""
-        skill_file_path = str(self.temp_dir.joinpath('.mycroft_skills'))
-        patch_path = (
-            self.mock_package +
-            'SkillUpdater.installed_skills_file_path'
-        )
+        skill_file_path = path.join(self.temp_dir, '.mycroft_skills')
+        patch_path = MOCK_PACKAGE + 'SkillUpdater.installed_skills_file_path'
         with patch(patch_path, new_callable=PropertyMock) as mock_file:
             mock_file.return_value = skill_file_path
-            updater = SkillUpdater(self.message_bus_mock)
+            updater = SkillUpdater(self.message_bus)
             updater.installed_skills = ['FooSkill', 'BarSkill']
             updater._save_installed_skills()
 
@@ -127,11 +223,11 @@ class TestSkillUpdater(MycroftUnitTestBase):
 
     def test_installed_skills_path_virtual_env(self):
         """Test the property representing the installed skill file path."""
-        with patch(self.mock_package + 'sys', spec=True) as sys_mock:
+        with patch(MOCK_PACKAGE + 'sys', spec=True) as sys_mock:
             sys_mock.executable = 'path/to/the/virtual_env/bin/python'
-            with patch(self.mock_package + 'os.access') as os_patch:
+            with patch(MOCK_PACKAGE + 'os.access') as os_patch:
                 os_patch.return_value = True
-                updater = SkillUpdater(self.message_bus_mock)
+                updater = SkillUpdater(self.message_bus)
                 self.assertEqual(
                     'path/to/the/virtual_env/.mycroft-skills',
                     updater.installed_skills_file_path
@@ -139,9 +235,9 @@ class TestSkillUpdater(MycroftUnitTestBase):
 
     def test_installed_skills_path_not_virtual_env(self):
         """Test the property representing the installed skill file path."""
-        with patch(self.mock_package + 'os.access') as os_patch:
+        with patch(MOCK_PACKAGE + 'os.access') as os_patch:
             os_patch.return_value = False
-            updater = SkillUpdater(self.message_bus_mock)
+            updater = SkillUpdater(self.message_bus)
             self.assertEqual(
                 path.expanduser('~/.mycroft/.mycroft-skills'),
                 updater.installed_skills_file_path
@@ -149,44 +245,48 @@ class TestSkillUpdater(MycroftUnitTestBase):
 
     def test_default_skill_names(self):
         """Test the property representing the list of default skills."""
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         self.assertIn('time', updater.default_skill_names)
         self.assertIn('weather', updater.default_skill_names)
         self.assertIn('test_skill', updater.default_skill_names)
 
     def test_download_skills_not_connected(self):
         """Test the error that occurs when the device is not connected."""
-        with patch(self.mock_package + 'connected') as connected_mock:
+        with patch(MOCK_PACKAGE + 'connected') as connected_mock:
             connected_mock.return_value = False
-            with patch(self.mock_package + 'time', spec=True) as time_mock:
+            with patch(MOCK_PACKAGE + 'time', spec=True) as time_mock:
                 time_mock.return_value = 100
-                updater = SkillUpdater(self.message_bus_mock)
-                result = updater.update_skills()
+                with patch(MOCK_PACKAGE + 'dialog.get') as dialog_mock:
+                    sm = SkillUpdater(self.message_bus)
+                    result = sm.download_skills(speak=True)
+                    dialog_mock.assert_called_once_with(
+                        'not connected to the internet'
+                    )
 
         self.assertFalse(result)
-        self.assertEqual(400, updater.next_download)
+        self.assertListEqual(self.message_bus.message_types, ['speak'])
+        self.assertEqual(400, sm.next_download)
 
     def test_post_manifest_allowed(self):
         """Test calling the skill manifest API endpoint"""
         self.msm_mock.skills_data = 'foo'
-        with patch(self.mock_package + 'is_paired') as paired_mock:
+        with patch(MOCK_PACKAGE + 'is_paired') as paired_mock:
             paired_mock.return_value = True
-            with patch(self.mock_package + 'DeviceApi', spec=True) as api_mock:
-                SkillUpdater(self.message_bus_mock).post_manifest()
+            with patch(MOCK_PACKAGE + 'DeviceApi', spec=True) as api_mock:
+                SkillUpdater(self.message_bus).post_manifest()
                 api_instance = api_mock.return_value
                 api_instance.upload_skills_data.assert_called_once_with('foo')
-            print(paired_mock.called)
-            paired_mock.assert_called_once_with()
+            paired_mock.assert_called_once()
 
     def test_get_skill_data(self):
         """Test invoking MSM to retrieve skill data."""
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         skill_data = updater._get_skill_data('test_skill')
         self.assertDictEqual(dict(name='test_skill', beta=False), skill_data)
 
     def test_get_skill_data_not_found(self):
         """Test invoking MSM to retrieve unknown skill data."""
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         skill_data = updater._get_skill_data('foo')
         self.assertDictEqual({}, skill_data)
 
@@ -195,7 +295,7 @@ class TestSkillUpdater(MycroftUnitTestBase):
         self.msm_mock.skills_data['skills'][0]['beta'] = True
         skill = self._build_mock_msm_skill_list()
         skill.is_local = False
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater.install_or_update(skill)
         self.assertIn('foobar', updater.installed_skills)
         self.assertIsNone(skill.sha)
@@ -203,11 +303,11 @@ class TestSkillUpdater(MycroftUnitTestBase):
     def test_install_or_update_local(self):
         """Test calling install_or_update with a local skill"""
         skill = self._build_mock_msm_skill_list()
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater.install_or_update(skill)
         self.assertIn('foobar', updater.installed_skills)
-        skill.update.assert_called_once_with()
-        skill.update_deps.assert_called_once_with()
+        skill.update.assert_called_once()
+        skill.update_deps.assert_called_once()
         self.msm_mock.install.assert_not_called()
 
     def test_install_or_update_default(self):
@@ -215,10 +315,10 @@ class TestSkillUpdater(MycroftUnitTestBase):
         skill = self._build_mock_msm_skill_list()
         skill.name = 'test_skill'
         skill.is_local = False
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater.install_or_update(skill)
         self.assertIn('test_skill', updater.installed_skills)
-        self.assertTrue(not skill.update.called)
+        skill.update.assert_not_once()
         self.msm_mock.install.assert_called_once_with(skill, origin='default')
 
     def test_install_or_update_default_fail(self):
@@ -227,11 +327,11 @@ class TestSkillUpdater(MycroftUnitTestBase):
         skill.name = 'test_skill'
         skill.is_local = False
         self.msm_mock.install.side_effect = ValueError
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         with self.assertRaises(ValueError):
             updater.install_or_update(skill)
         self.assertNotIn('test_skill', updater.installed_skills)
-        self.assertTrue(not skill.update.called)
+        skill.update.assert_not_once()
         self.msm_mock.install.assert_called_once_with(skill, origin='default')
         self.assertTrue(updater.default_skill_install_error)
 
@@ -244,13 +344,19 @@ class TestSkillUpdater(MycroftUnitTestBase):
         skill.install = Mock()
         skill.update = Mock()
         skill.update_deps = Mock()
-        skill.path = str(self.temp_dir.joinpath('foobar'))
+        skill.path = path.join(self.temp_dir, 'foobar')
 
         return skill
 
+    def test_speak_skill_updated(self):
+        """Test emitting a speak event to the bus when skills are updated."""
+        SkillUpdater(self.message_bus)._speak_skill_updated(speak=True)
+        self.dialog_mock.assert_called_once_with('skills updated')
+        self.assertListEqual(['speak'], self.message_bus.message_types)
+
     def test_schedule_retry(self):
         """Test scheduling a retry of a failed install."""
-        updater = SkillUpdater(self.message_bus_mock)
+        updater = SkillUpdater(self.message_bus)
         updater._schedule_retry()
         self.assertEqual(1, updater.install_retries)
         self.assertEqual(400, updater.next_download)
@@ -258,10 +364,10 @@ class TestSkillUpdater(MycroftUnitTestBase):
 
     def test_update_download_time(self):
         """Test updating the next time a download will occur."""
-        dot_msm_path = self.temp_dir.joinpath('.msm')
-        dot_msm_path.touch()
-        dot_msm_mtime_before = dot_msm_path.stat().st_mtime
+        dot_msm_path = path.join(self.temp_dir, '.msm')
+        Path(dot_msm_path).touch()
+        dot_msm_mtime_before = path.getmtime(dot_msm_path)
         sleep(0.5)
-        SkillUpdater(self.message_bus_mock)._update_download_time()
-        dot_msm_mtime_after = dot_msm_path.stat().st_mtime
+        SkillUpdater(self.message_bus)._update_download_time()
+        dot_msm_mtime_after = path.getmtime(path.join(self.temp_dir, '.msm'))
         self.assertLess(dot_msm_mtime_before, dot_msm_mtime_after)
